THRIFT-916 long long becomes int64_t
Patch: Christian Lavoie



git-svn-id: https://svn.apache.org/repos/asf/thrift/trunk@1037500 13f79535-47bb-0310-9956-ffa450edef68

+++ contrib/fb303/TClientInfo.cpp
@@ -160,7 +160,7 @@ void TClientInfoServerHandler::getStatsStrings(vector<string>& result) {
 
     char buf[256];
     snprintf(buf, sizeof buf, "%d %s %s %.3f %llu", i, addrStr, callStr, secs,
-             (unsigned long long)info->getNCalls());
+             (uint64_t)info->getNCalls());
                
     result.push_back(buf);
   }
